+++ packages/bundle/src/markdown/middleware/createCodeBlockCopyButtonMiddleware.ts
@@ -7,6 +7,16 @@ export default function createCodeBlockCopyButtonMiddleware(): HTMLContentTransf
     next(
       Object.freeze({
         ...request,
+        allowedTags: Object.freeze(
+          new Map(request.allowedTags).set(
+            request.codeBlockCopyButtonTagName,
+            Object.freeze({
+              attributes: Object.freeze(
+                new Set(['class', 'data-alt-copy', 'data-alt-copied', 'data-testid', 'data-value'])
+              )
+            })
+          )
+        ),
         documentFragment: codeBlockCopyButtonDocumentMod(request.documentFragment, {
           codeBlockCopyButtonAltCopied: request.codeBlockCopyButtonAltCopied,
           codeBlockCopyButtonAltCopy: request.codeBlockCopyButtonAltCopy,